pyproject.toml
[tool.poetry]
name = "index-suep-channel"
version = "0.2.0"
description = ""
authors = []
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.8.5"
argparse = "^1.4.0"
htmlmin = "^0.1.12"
jinja2 = "^3.1.2"
markupsafe = "^2.1.3"
paddleocr = "^2.7.0.3"
paddlepaddle = "^2.5.1"
pillow = "^10.0.1"
toml = "^0.10.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
flake8 = "^6.1.0"

[tool.poetry.group.ci]
optional = true

[tool.poetry.group.ci.dependencies]
selenium = "^4.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
